refactor: sealed class 파일 내부에 정리
