+++ app/src/main/java/woowacourse/shopping/data/model/coupon/AvailableTimeResponse.kt
@@ -1,15 +1,11 @@
 package woowacourse.shopping.data.model.coupon
 
-import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonClassDiscriminator
 import woowacourse.shopping.domain.model.AvailableTime
 import java.time.LocalTime
 
-@OptIn(ExperimentalSerializationApi::class)
 @Serializable
-@JsonClassDiscriminator("discountType")
 data class AvailableTimeResponse(
     @SerialName("end")
     val end: String,
